Added L8 - Forms assignment files
FormsAssignment.html and validationScript.js are an assignment from the Web-115  Web Markup and Scripting at wake tech. Uploading to git hub is a mini assignment to become more familiar with git and github. The Lesson 8 forms assignment was used to teach and provide practice validating forms with JavaScript. The  assignment involved validating a forming that includes a single input field for alphanumeric values. regex was used to validate the input value and provide feedback to the user based on whether the input matches the required pattern.